requirements.txt
anyio==3.6.2
cachetools==5.2.0
certifi==2022.12.7
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
Pillow==9.4.0
psycopg2==2.9.5
python-telegram-bot==20.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.0rc1
tabulate==0.9.0
typing_extensions==4.4.0
